move all indexing tests to algorithms/indexing
